Added outline for Backend game logic
